# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This api call data from youtube anf table",
        "title": "Dataset Collection API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.201.252.7:30014",
    "paths": {
...
        "/dataset/caracdnt/{carAccidentId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataset car accident"
                ],
                "summary": "get list by carAccidentId",
                "operationId": "query-carAccident-by-id-handler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Car Accident Id",
                        "name": "carAccidentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need CarAccidentId!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
    }
}